+++ src/Mscc.GenerativeAI/VertexAI.cs
@@ -58,14 +58,14 @@ namespace Mscc.GenerativeAI
         /// <summary>
         /// Create a generative model on Vertex AI to use.
         /// </summary>
-        /// <param name="model">Model to use (default: "gemini-1.0-pro")</param>
+        /// <param name="model">Model to use (default: "gemini-1.5-pro")</param>
         /// <param name="generationConfig">Optional. Configuration options for model generation and outputs.</param>
         /// <param name="safetySettings">Optional. A list of unique SafetySetting instances for blocking unsafe content.</param>
         /// <param name="tools">Optional. A list of Tools the model may use to generate the next response.</param>
         /// <param name="systemInstruction">Optional. </param>
         /// <exception cref="ArgumentNullException">Thrown when <paramref name="projectId"/> or <paramref name="region"/> is <see langword="null"/>.</exception>
         /// <returns>Generative model instance.</returns>
-        public GenerativeModel GenerativeModel(string model = Model.Gemini10Pro,
+        public GenerativeModel GenerativeModel(string model = Model.Gemini15Pro,
             GenerationConfig? generationConfig = null,
             List<SafetySetting>? safetySettings = null,
             List<Tool>? tools = null,